+++ src/Decorators/MobileAppUrlDecorator.php
@@ -19,6 +19,13 @@ class MobileAppUrlDecorator extends ModeDecoratorBase
                 $content['mobile_app_url'] = url()->route('mobile.pack.lesson.show', [$content['id']]);
             } elseif ($content['type'] == 'semester-pack-lesson') {
                 $content['mobile_app_url'] = url()->route('mobile.pack.lesson.show', [$content['id']]);
+            }elseif($content['type'] == 'learning-path-level'){
+                $content['banner_button_url'] = $content->fetch('current_lesson') ? url()->route(
+                    'mobile.learning-path.lesson.show',
+                    [
+                        $content->fetch('current_lesson')['id'],
+                    ]
+                ) : '';
             } elseif ($content['type'] == 'learning-path-course') {
                 $content['mobile_app_url'] = url()->route(
                     'mobile.learning-path.course.show',
@@ -26,6 +33,13 @@ class MobileAppUrlDecorator extends ModeDecoratorBase
                         $content['id'],
                     ]
                 );
+
+                $content['banner_button_url'] = $content->fetch('current_lesson') ? url()->route(
+                    'mobile.learning-path.lesson.show',
+                    [
+                        $content->fetch('current_lesson')['id'],
+                    ]
+                ) : '';
             } elseif($content['type'] == 'learning-path-lesson'){
                 $content['mobile_app_url'] =url()->route(
                     'mobile.learning-path.lesson.show',
